requirements.txt
pyotp~=2.6.0
typer~=0.4.0
pyqrcode~=1.2.1
pyzbar~=0.1.8
pillow~=8.4.0
click~=8.0.3